requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
bitarray==3.0.0
Brownie==0.5.1
cachetools==5.5.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
cycler==0.12.1
cytoolz==1.0.1
Django==5.1.5
django-allauth==65.4.0
djangorestframework==3.15.2
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.2.0
ffmpeg-python==0.2.0
fonttools==4.55.3
frozenlist==1.5.0
future==1.0.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
kiwisolver==1.4.8
kombu==5.4.2
lazy_loader==0.4
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
parsimonious==0.10.0
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
pytz==2024.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.0.1
rsa==4.9
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tifffile==2025.1.10
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
web3==7.7.0
websockets==13.1
yarl==1.18.3
python-dotenv==1.0.1
django-celery-beat
djangorestframework
redis
billiard
werkzeug
gunicorn
whitenoise